Rename serializers to rpcs.py
